Fix error on decoding of an empty list in an SSZ Union (#73)
This removes a check that was (probably) there to verify if the
input data had more than just the selector byte. However, there
is the case of an empty SSZ.List that encoded as 0 bytes.

This does sound a bit strange as the potential None value in
an SSZ Union also has 0 bytes, but both cases can be decoded
properply as a None is defined by having no field (which is
checked at compile time).