Avoid pointless nim-faststreams copies
When faststreams doesn't know the size of the buffer, it'll keep
allocating small temporary buffers (pages) where data is kept until it's
gathered / copied into a single contiguous buffer at the end of the
write.

This is doubly inefficient, not only because all data is copied twice
but it's also zero:ed twice due to a fundamental inefficiency in
faststream:s where data is held in a `string` and therefore
zero-initialized.

By pre-allocating the buffer, we avoid all page handling in faststreams
including the allocation of many small buffers - also when calling
`getOutput` - and we only have to pay one redundant zeroing of the
buffer.