Implement EIP-8016: SSZ CompatibleUnion
Add support for `CompatibleUnion` as specced out in EIP-8016, without
the type checker from the Python implementation:

- https://eips.ethereum.org/EIPS/eip-8016

This remains compatible with the regular `Union` used by Fluffy.
`CompatibleUnion` uses the same serialization / merkleization;
it restricts type options to have a shared compatible tree shape
(which we don't need to check as Python already does it), and it
allows holey enums that don't start at 0 for forward compatibility.

Python implementation:

- https://github.com/ethereum/remerkleable/pull/5